feat: Added support for Message description (#211)
While working at https://github.com/springwolf/springwolf-core/pull/189 we found out that the provided Message description was, in fact, the one associated to the Operation. As a result we decided to fully remove the description until we could fix it.

This PR adds support for Message description.

Any `Message` object, annotated with `@Schema` will use the annotation `description` value, as the Message description.