feat: Added @AsyncMessage (#231)
* feat: Added @AsyncMessage

Added support for a new `@AsyncMessage` annotation. This annotation can be used as a method parameter annotation on the same method the `@AsyncPublisher` or `@AsyncListener` annotations are used.

This new `@AsyncMessage` annotation allows to enrich the AsyncAPI Operation Message with all the supported fields.

The Message payload documentation stays in the Swagger `@Schema` annotation. This new `@AsyncMessage` annotation is intended to enrich the missing values.

A special case is the description, which is available in both. Any value defined in `@AsyncMessage` has higher priority than any other default value.

* feat: Define @AsyncMessage as part of @AsyncOperation

Following the spec https://www.asyncapi.com/docs/reference/specification/v2.6.0#operationObject the `message` value should be part of the operationObject.

* Apply suggestions from code review

---------

Co-authored-by: Timon Back <timonback@users.noreply.github.com>