SpringwolfInitApplicationListener: moved fail-fast path from 'afterPropertiesSet' to onApplicationReadyEvent (#431)
This PR introduces a small change to the fail-fast initialization of AsyncApi. Until now, the initialization happend during afterPropertiesSet() method in SpringWolfInitApplicationListener - which means: after the SpringWolfInitApplicationListener bean and all dependents (including AsyncApiService etc...) have been initialized.

The disadvantage of initializing the AsyncAPI at this point of time is, that the spring context isn't completly initialized, some beans may not exist in the spring-context at this moment. This may be problematic if scanner classes will for example make use of BeanPostProcessors or Spring AOP to detect (infrastructre) beans.

I moved the fail-fast initializing branch into the onApplicationEvent() method, which means:

the AsyncApi initialization happens a little bit later
the fail-fast semantics remains valid, in case of a validation error the spring application will not boot but abort throwing an exception.