chore: Improved AsyncAPI interface (#515)
* chore: Improved AsyncAPI interface

Improved the creation of Messages.

A 'Message' is now an interface that's implemented by
'MessageObject' and 'MessageReference'. This change
allows to properly define a 'Message' and use it in a more
transparent way.

Added also different builders

* chore: Improved AsyncAPI interface

Improved the creation of Schemas.

A 'Schemas' is now an interface that's implemented by
'SchemasObject' and 'SchemasReference'. This change
allows to properly define a 'Schemas' and use it in a more
transparent way.

Added also different builders

* chore: Improved AsyncAPI interface

Improved the creation of Messages.

A 'Message' is now an interface that's implemented by
'MessageObject' and 'MessageReference'. This change
allows to properly define a 'Message' and use it in a more
transparent way.

Added also different builders

* chore: Move 'net.javacrumbs.json-unit:json-unit-assertj' version to dependencies.gradle

* chore: Improved AsyncAPI enums

Added convenient methods to properly create the enums from their
String representations.

* fix: MessageBinding should be abstract

* feat: Added missing SNS Binding Operation

feat: Added missing SNS Binding Operation

* fix: Fixed Operation Bindings

fix: Fixed spotless