fix: Fixed core tests (#540)
* fix: Fixed core tests

Fixed the failing tests after migrating to AsyncAPI v3

* fix: Fixed AMQP integration

The AMQP was adding both 'exchange' & 'queue' keys, but it should only include the
one related with the binding type.

Also, deliveryMode can be ONLY 1 or 2, the
default value 0 which was used in different places is wrong.

* fix: Fixed AMQP integration

The AMQP was adding both 'exchange' & 'queue' keys, but it should only include the
one related with the binding type.

Also, deliveryMode can be ONLY 1 or 2, the
default value 0 which was used in different places is wrong.