(fix) Fixed a bug loading the Kotlinx Module without `use-fqn` variable
When the Kotlinx Serialization Module was available in an environment without the `springwolf.use-fqn` the code was failing, since it was trying to inject the variable.

The code is now loading the converter **only** if `springwolf.enabled` is `true` (default value). The `springwolf.use-fqn` is also defaulting to `true`.

This solution fixes the identified issue.